+++ filecoin-proofs/src/api/sector_builder/kv_store/fs.rs
@@ -67,3 +67,25 @@ impl KeyValueStore for FileSystemKvs {
         }
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_alpha() {
+        let metadata_dir = tempfile::tempdir().unwrap();
+        let db = FileSystemKvs::initialize(metadata_dir).unwrap();
+
+        let k_a = b"key-xx";
+        let k_b = b"key-yy";
+        let v_a = b"value-aa";
+        let v_b = b"value-bb";
+
+        db.put(k_a, v_a).unwrap();
+        db.put(k_b, v_b).unwrap();
+
+        let opt = db.get(k_a).unwrap();
+        assert_eq!(format!("{:x?}", opt.unwrap()), format!("{:x?}", v_a));
+    }
+}
